+++ packages/@expo/cli/src/start/server/metro/MetroBundlerDevServer.ts
@@ -93,6 +93,7 @@ import {
   convertPathToModuleSpecifier,
   createBundleUrlPath,
   createBundleOsPath,
+  getAssetPrefixFromExpoConfig,
   getAsyncRoutesFromExpoConfig,
   getBaseUrlFromExpoConfig,
   getMetroDirectBundleOptions,
@@ -674,8 +675,16 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     request?: ImmutableRequest
   ): Promise<{ content: string | ReadableStream<Uint8Array>; resources?: SerialAsset[] }> {
     const { exp } = getConfig(this.projectRoot);
-    const { mode, isExporting, clientBoundaries, baseUrl, reactCompiler, routerRoot, asyncRoutes } =
-      this.instanceMetroOptions;
+    const {
+      mode,
+      isExporting,
+      clientBoundaries,
+      baseUrl,
+      assetPrefix,
+      reactCompiler,
+      routerRoot,
+      asyncRoutes,
+    } = this.instanceMetroOptions;
     assert(
       mode != null &&
         isExporting != null &&
@@ -783,6 +792,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
       template: staticHtml,
       devBundleUrl: devBundleUrlPathname,
       baseUrl,
+      assetPrefix,
       hydrate: env.EXPO_WEB_DEV_HYDRATE,
     });
     return {
@@ -1280,6 +1290,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     const useServerRendering = ['static', 'server'].includes(exp.web?.output ?? '');
     const hasApiRoutes = isReactServerComponentsEnabled || exp.web?.output === 'server';
     const baseUrl = getBaseUrlFromExpoConfig(exp);
+    const assetPrefix = getAssetPrefixFromExpoConfig(exp) || undefined;
     const asyncRoutes = getAsyncRoutesFromExpoConfig(exp, options.mode ?? 'development', 'web');
     const routerRoot = getRouterDirectoryModuleIdWithManifest(this.projectRoot, exp);
     const reactCompiler = !!exp.experiments?.reactCompiler;
@@ -1306,6 +1317,7 @@ export class MetroBundlerDevServer extends BundlerDevServer {
     const instanceMetroOptions = {
       isExporting: !!options.isExporting,
       baseUrl,
+      assetPrefix,
       mode,
       routerRoot,
       reactCompiler,