+++ packages/@expo/cli/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🎉 New features
 
 - Support Bundler-managed CocoaPods installations ([#43605](https://github.com/expo/expo/pull/43605) by [@tiwari91](https://github.com/tiwari91), [@kitten](https://github.com/kitten))
+- Add `experiments.assetPrefix` config option to serve static assets from a CDN while keeping the app URL unchanged. ([#41520](https://github.com/expo/expo/pull/41520) by [@jtomaszewski](https://github.com/jtomaszewski))
 
 ### 🐛 Bug fixes